requirements.txt
django===4.1.3
pillow
gunicorn
whitenoise
django-environ
django-storages
boto3
django-admin-csvexport
pycryptodome
pay_ccavenue
shortuuid
sendgrid
pytz